pyproject.toml
[project]
name = "rocketinst-timetable-uploader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "img2table>=1.4.2",
    "ipykernel>=7.1.0",
    "pandas>=2.3.3",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.2",
    "torch>=2.9.1",
    "yomitoku>=0.5.3",
]
